reorg: move scan to core.scan
