+++ ggshield/cmd/sca/scan/sca_scan_utils.py
@@ -11,9 +11,9 @@ from ggshield.core.config.user_config import SCAConfig
 from ggshield.core.errors import APIKeyCheckError, UnexpectedError
 from ggshield.core.file_utils import get_empty_tar
 from ggshield.core.git_shell import INDEX_REF, tar_from_ref_and_filepaths
+from ggshield.core.scan.scan_context import ScanContext
+from ggshield.core.scan.scan_mode import ScanMode
 from ggshield.core.text_utils import display_error, display_info, display_warning
-from ggshield.scan.scan_context import ScanContext
-from ggshield.scan.scan_mode import ScanMode
 from ggshield.verticals.sca.client import SCAClient
 from ggshield.verticals.sca.file_selection import (
     get_all_files_from_sca_paths,
