+++ ggshield/cmd/secret/scan/archive.py
@@ -11,8 +11,8 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 )
 from ggshield.core.config import Config
 from ggshield.core.errors import UnexpectedError
-from ggshield.scan import Files, ScanContext, ScanMode
-from ggshield.scan.file import get_files_from_paths
+from ggshield.core.scan import Files, ScanContext, ScanMode
+from ggshield.core.scan.file import get_files_from_paths
 from ggshield.verticals.secret import (
     RichSecretScannerUI,
     SecretScanCollection,
