+++ ggshield/core/scan/__init__.py
