+++ ggshield/core/scan/commit.py
