+++ ggshield/core/scan/file.py
