+++ ggshield/core/scan/id_cache.py
