+++ ggshield/core/scan/scan_context.py
