+++ ggshield/core/scan/scan_mode.py
