+++ ggshield/core/scan/scannable.py
