+++ ggshield/verticals/sca/file_selection.py
@@ -11,8 +11,8 @@ from ggshield.core.git_shell import (
     get_filepaths_from_ref,
     get_staged_filepaths,
 )
-from ggshield.scan import Scannable
-from ggshield.scan.file import get_files_from_paths
+from ggshield.core.scan import Scannable
+from ggshield.core.scan.file import get_files_from_paths
 from ggshield.verticals.sca.client import SCAClient
