+++ tests/unit/cmd/scan/test_prereceive.py
@@ -5,8 +5,8 @@ from click.testing import CliRunner
 
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
+from ggshield.core.scan import StringScannable
 from ggshield.core.utils import EMPTY_SHA, Filemode
-from ggshield.scan import StringScannable
 from ggshield.verticals.secret import Result, Results, SecretScanCollection
 from ggshield.verticals.secret.repo import cd
 from tests.repository import Repository, create_pre_receive_repo
