+++ tests/unit/cmd/test_ignore.py
@@ -6,8 +6,8 @@ from pygitguardian.models import Match, PolicyBreak
 from ggshield.cmd.secret.ignore import ignore_last_found
 from ggshield.core.cache import Cache
 from ggshield.core.config import Config
+from ggshield.core.scan import Commit, ScanContext, ScanMode
 from ggshield.core.types import IgnoredMatch
-from ggshield.scan import Commit, ScanContext, ScanMode
 from ggshield.verticals.secret import SecretScanner
 from tests.unit.conftest import _MULTIPLE_SECRETS_PATCH, my_vcr
