+++ tests/unit/core/scan/snapshots/snap_test_scannable.py
