+++ tests/unit/verticals/sca/test_file_selection.py
@@ -5,8 +5,8 @@ from typing import Set
 import pytest
 from pygitguardian import GGClient
 
+from ggshield.core.scan import StringScannable
 from ggshield.core.utils import Filemode
-from ggshield.scan import StringScannable
 from ggshield.verticals.sca.client import SCAClient
 from ggshield.verticals.sca.file_selection import (
     get_all_files_from_sca_paths,
