+++ tests/unit/verticals/secret/output/test_json_output.py
@@ -9,7 +9,7 @@ from voluptuous import Optional, Required, validators
 
 from ggshield.core.filter import leak_dictionary_by_ignore_sha
 from ggshield.core.git_shell import Filemode
-from ggshield.scan import Commit, ScanContext, ScanMode, StringScannable
+from ggshield.core.scan import Commit, ScanContext, ScanMode, StringScannable
 from ggshield.verticals.secret import (
     Result,
     Results,
