+++ tests/unit/verticals/secret/output/test_text_output.py
@@ -6,7 +6,7 @@ import pytest
 
 from ggshield.core.filter import leak_dictionary_by_ignore_sha
 from ggshield.core.git_shell import Filemode
-from ggshield.scan import StringScannable
+from ggshield.core.scan import StringScannable
 from ggshield.verticals.secret import Result, Results, SecretScanCollection
 from ggshield.verticals.secret.output import SecretTextOutputHandler
 from ggshield.verticals.secret.output.secret_text_output_handler import (
